move LoggerRt into Jewel
